+++ examples-pydantic-v1/windmill_pydantic_v1/data_classes/_high_speed_shaft.py
@@ -61,9 +61,9 @@ class HighSpeedShaftGraphQL(GraphQLCore):
     """
 
     view_id = dm.ViewId("power-models", "HighSpeedShaft", "1")
-    bending_moment_y: Union[TimeSeries, str, None] = None
-    bending_monent_x: Union[TimeSeries, str, None] = None
-    torque: Union[TimeSeries, str, None] = None
+    bending_moment_y: Union[TimeSeries, dict, None] = None
+    bending_monent_x: Union[TimeSeries, dict, None] = None
+    torque: Union[TimeSeries, dict, None] = None
 
     @root_validator(pre=True)
     def parse_data_record(cls, values: Any) -> Any:
@@ -75,6 +75,14 @@ class HighSpeedShaftGraphQL(GraphQLCore):
                 last_updated_time=values.pop("lastUpdatedTime", None),
             )
         return values
+
+    @validator("bending_moment_y", "bending_monent_x", "torque", pre=True)
+    def parse_timeseries(cls, value: Any) -> Any:
+        if isinstance(value, list):
+            return [TimeSeries.load(v) if isinstance(v, dict) else v for v in value]
+        elif isinstance(value, dict):
+            return TimeSeries.load(value)
+        return value
 
     def as_read(self) -> HighSpeedShaft:
         """Convert this GraphQL format of high speed shaft to the reading format."""