+++ examples-pydantic-v1/windmill_pydantic_v1/data_classes/_power_inverter.py
@@ -61,9 +61,9 @@ class PowerInverterGraphQL(GraphQLCore):
     """
 
     view_id = dm.ViewId("power-models", "PowerInverter", "1")
-    active_power_total: Union[TimeSeries, str, None] = None
-    apparent_power_total: Union[TimeSeries, str, None] = None
-    reactive_power_total: Union[TimeSeries, str, None] = None
+    active_power_total: Union[TimeSeries, dict, None] = None
+    apparent_power_total: Union[TimeSeries, dict, None] = None
+    reactive_power_total: Union[TimeSeries, dict, None] = None
 
     @root_validator(pre=True)
     def parse_data_record(cls, values: Any) -> Any:
@@ -75,6 +75,14 @@ class PowerInverterGraphQL(GraphQLCore):
                 last_updated_time=values.pop("lastUpdatedTime", None),
             )
         return values
+
+    @validator("active_power_total", "apparent_power_total", "reactive_power_total", pre=True)
+    def parse_timeseries(cls, value: Any) -> Any:
+        if isinstance(value, list):
+            return [TimeSeries.load(v) if isinstance(v, dict) else v for v in value]
+        elif isinstance(value, dict):
+            return TimeSeries.load(value)
+        return value
 
     def as_read(self) -> PowerInverter:
         """Convert this GraphQL format of power inverter to the reading format."""