+++ examples/windmill/data_classes/_sensor_position.py
@@ -93,14 +93,14 @@ class SensorPositionGraphQL(GraphQLCore):
     """
 
     view_id = dm.ViewId("power-models", "SensorPosition", "1")
-    edgewise_bend_mom_crosstalk_corrected: Union[TimeSeries, str, None] = None
-    edgewise_bend_mom_offset: Union[TimeSeries, str, None] = None
-    edgewise_bend_mom_offset_crosstalk_corrected: Union[TimeSeries, str, None] = None
-    edgewisewise_bend_mom: Union[TimeSeries, str, None] = None
-    flapwise_bend_mom: Union[TimeSeries, str, None] = None
-    flapwise_bend_mom_crosstalk_corrected: Union[TimeSeries, str, None] = None
-    flapwise_bend_mom_offset: Union[TimeSeries, str, None] = None
-    flapwise_bend_mom_offset_crosstalk_corrected: Union[TimeSeries, str, None] = None
+    edgewise_bend_mom_crosstalk_corrected: Union[TimeSeries, dict, None] = None
+    edgewise_bend_mom_offset: Union[TimeSeries, dict, None] = None
+    edgewise_bend_mom_offset_crosstalk_corrected: Union[TimeSeries, dict, None] = None
+    edgewisewise_bend_mom: Union[TimeSeries, dict, None] = None
+    flapwise_bend_mom: Union[TimeSeries, dict, None] = None
+    flapwise_bend_mom_crosstalk_corrected: Union[TimeSeries, dict, None] = None
+    flapwise_bend_mom_offset: Union[TimeSeries, dict, None] = None
+    flapwise_bend_mom_offset_crosstalk_corrected: Union[TimeSeries, dict, None] = None
     position: Optional[float] = None
 
     @model_validator(mode="before")