+++ tests/test_integration/test_graphql_query.py
@@ -1,11 +1,17 @@
 from __future__ import annotations
 
+from cognite.client.data_classes import TimeSeries
+
 from tests.constants import IS_PYDANTIC_V2
 
 if IS_PYDANTIC_V2:
+    from omni import OmniClient
+    from omni import data_classes as odc
     from windmill import WindmillClient
     from windmill import data_classes as wdc
 else:
+    from omni_pydantic_v1 import OmniClient
+    from omni_pydantic_v1 import data_classes as odc
     from windmill_pydantic_v1 import WindmillClient
     from windmill_pydantic_v1 import data_classes as wdc
 
@@ -24,3 +30,34 @@ def test_graphql_query(wind_client: WindmillClient) -> None:
     assert len(result) == 1
     assert isinstance(result[0], wdc.BladeGraphQL)
     assert result[0].name
+
+
+def test_query_cdf_external(omni_client: OmniClient) -> None:
+    query = """{
+  listCDFExternalReferencesListed{
+    items{
+       __typename
+      timeseries {
+        externalId
+        name
+      }
+      externalId
+      files{
+        externalId
+        name
+      }
+      sequences{
+        externalId
+        name
+
+      }
+    }
+  }
+}
+"""
+    result = omni_client.graphql_query(query)
+
+    assert len(result) > 0
+    assert isinstance(result[0], odc.CDFExternalReferencesListedGraphQL)
+    assert len(result[0].timeseries) > 0
+    assert isinstance(result[0].timeseries[0], TimeSeries)