+++ docs/CHANGELOG.md
@@ -13,6 +13,11 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Fixed
+- When retrieving `.graphql` with a CDF external resource type (`Timeseries`, `File`, or `Sequence`), the generated SDK
+  would raise a `ValidationError`. This is now fixed.
+
 ## [0.99.18] - 03-04-24
 ### Changed
 * Typing extensions, `typing-extensions` is no longer a direct dependency of `pygen`. This is to avoid